Move switchLinkType to state/actions
